+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 ## [Unreleased]
 ### Changed
 - Moved tests out to test directory and reorganised
+- Simplified the Hamming tests
 - Tidied up documentation
 
 ## [0.8.0] - (2018-08-19)